Display last line of a long message + DPAD scrolling
If the text is smaller than the screen, it retains its normal behavior (top/middle/bottom alignment).
If text is larger than the screen, it is automatically shifted upwards so that the last line is visible
Lines that leave the screen at the top are not rendered (optimization).
DPAD scrolling so tsee previous messages